backend/requirements.txt
google-adk
google-genai
fastapi
uvicorn[standard]
pydantic
pydantic-settings
sse-starlette
python-dotenv
httpx
